feat: substitute downward then jump to the top for the rest
The substitute command with % as the pattern is really 1,$. Which is
first line to the last line. So when the substitution starts, if there
are matches above the cursor line the cursor will jump up to the first
match and start the substituting downward.

This feels jarring. So let's try to address that by starting at the
current position down. Then repeating the last substitute with a new
range of 1st line to the line above the cursor position.

This is a lot of code for this nicety. Time will tell if it's worth it.